--- converter/package.json
@@ -1,10 +1,0 @@
-{
-  "private": true,
-  "scripts": {
-    "dev": "vitest dev"
-  },
-  "dependencies": {
-    "magic-regexp": "latest",
-    "regexp-tree": "^0.1.27"
-  }
-}
