+++ docs/content/2.getting-started/5.converter.md
@@ -1,17 +1,11 @@
-# magic-regexp converter
+---
+title: Converter (experimental)
+---
 
-> [!IMPORTANT]
-> This package is under development.
-
-## Features
-
-- Transform existing Regex into magic-regexp
-- Track regex features that are currently not supported by magic-regexp
-
-## Usage
+It is also possible to convert existing regular expressions to `magic-regexp` syntax.
 
 ```ts
-import { convert } from '.'
+import { convert } from 'magic-regexp/converter'
 
 convert(/a|b|c/)
 // createRegExp(exactly('a').or('b').or('c'))
@@ -22,8 +16,8 @@ convert(/(foo)bar\d+/)
 
 ### Options
 
-- `argsOnly` (boolean)  
-  _Default: `false`_  
+- `argsOnly` (boolean)
+  _Default: `false`_
   Only show arguments without `createRegExp`
 
 ```ts
