+++ package.json
@@ -10,6 +10,10 @@
       "import": "./dist/index.mjs",
       "require": "./dist/index.cjs"
     },
+    "./converter": {
+      "import": "./dist/converter.mjs",
+      "require": "./dist/converter.cjs"
+    },
     "./transform": {
       "import": "./dist/transform.mjs",
       "require": "./dist/transform.cjs"
@@ -48,6 +52,7 @@
     "estree-walker": "^3.0.3",
     "magic-string": "^0.30.5",
     "mlly": "^1.4.2",
+    "regexp-tree": "^0.1.27",
     "type-level-regexp": "~0.1.17",
     "ufo": "^1.3.1",
     "unplugin": "^1.5.0"
