+++ pnpm-workspace.yaml
@@ -1,4 +1,3 @@
 packages:
   - "playground"
   - "docs"
-  - "converter"
